refactor: consolidate plugin error classifiers and merge display into output
- Add brutus.NewClassifier() helper that returns an error classifier
  function from auth indicators, replacing 13 identical 3-line wrapper
  functions across plugins with a single-line var declaration each
- Merge cmd/brutus/display.go into cmd/brutus/output.go since both
  handle display/formatting concerns (colors, symbols, banners, results)

Plugins with custom error classification logic (SSH, POP3, FTP, WinRM)
are left unchanged since their classifiers contain real logic beyond
simple delegation.
